requirements.txt
python-irodsclient~=0.8.2
psutil>=5.0.0
